requirements.txt
absl-py==0.7.1
antiorm==1.2.1
apparmor==2.10.95
apparmor.click==0.3.13.1
apturl==0.5.2
astor==0.8.0
astroid==1.6.3
attrs==19.3.0
audioread==2.1.8
autoenv==1.0.0
autopep8==1.5.2
autopilot==1.5.0
awal==0.0.0
bandmat==0.6
base==1.0.4
beautifulsoup4==4.4.1
bleach==1.5.0
blinker==1.3
bootstrapped==0.0.2
boto==2.49.0
boto3==1.7.78
botocore==1.10.78
bottle==0.12.17
Brlapi==0.6.4
bz2file==0.98
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
checkbox-support==0.22
cheroot==5.8.3
CherryPy==11.0.0
chrome-gnome-shell==0.0.0
Click==7.0
click-reviewers-tools===0.44-16.04.1
clickclick==1.2.2
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
configparser==4.0.2
conllu==0.4
connection==2019.4.13
connexion==2.4.0
cryptography==2.8
cycler==0.10.0
cymem==1.31.2
Cython==0.28.5
cytoolz==0.8.2
data==0.4
db==0.1.1
decorator==4.1.2
defer==1.0.6
descartes==1.1.0
dictset==0.3.1.2
dill==0.2.7.1
docopt==0.6.2
docutils==0.15.2
entrypoints==0.2.3
evdev==0.4.7
ExtensionClass==4.3.0
extras==0.0.3
fastdtw==0.3.2
feedparser==5.1.3
fixtures==1.3.1
Flask==1.1.1
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.1
fr-core-news-md==2.0.0
fr-core-news-sm==2.0.0
ftfy==4.4.3
funcsigs==1.0.2
gast==0.2.2
gensim==3.5.0
genson==1.2.1
glove-python==0.1.0
google-pasta==0.1.7
graphics==1.0
graphviz==0.8.4
grpcio==1.22.0
guacamole==0.9.2
h5py==2.9.0
hdbscan==0.8.24
html5lib==0.9999999
httplib2==0.9.1
idna==2.8
imbalanced-learn==0.3.3
imblearn==0.0
importlib-metadata==0.23
infinity==1.4
inflect==0.2.5
inflection==0.3.1
intervals==0.8.0
intervaltree==2.1.0
intonation==0.1
ipykernel==4.6.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.1
isort==4.3.4
itsdangerous==1.1.0
jedi==0.11.0
Jinja2==2.10.3
jmespath==0.9.3
joblib==0.13.2
jsonschema==3.1.1
junitxml==0.6
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.4.0
kaggle==1.5.6
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
language-selector==0.1
lazy-object-proxy==1.3.1
LibAppArmor==2.10.95
librosa==0.6.3
linecache2==1.0.0
livereload==2.4.0
llvmlite==0.29.0
louis==2.6.4
lpthw.web==1.1
lxml==3.5.0
Magic-file-extensions==0.2
Mako==1.0.3
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
mistune==0.8.3
mizani==0.5.4
mkdocs==0.14.0
more-itertools==7.2.0
msgpack-numpy==0.4.1
msgpack-python==0.4.8
mtcnn==0.1.0
multiproc==1.0.1
murmurhash==0.28.0
nbconvert==5.3.1
nbformat==4.4.0
nltk==3.2.5
nose==1.3.7
notebook==5.2.2
numba==0.44.1
numexpr==2.4.3
numpy==1.17.2
oauthlib==1.0.3
onboard==1.2.0
openapi-spec-validator==0.2.8
opencv-python==4.1.2.30
padasip==1.1.1
padme==1.1.1
palettable==3.3.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.1.0
pathlib==1.0.1
patsy==0.4.1
pbr==1.8.0
pexpect==4.0.1
pickleshare==0.7.4
Pillow==3.1.2
pipenv==11.10.0
pkginfo==1.4.1
plac==0.9.6
plainbox==0.25
plotnine==0.5.1
pocketsphinx==0.1.15
portend==2.1.2
pprint==0.1
praatio==3.6.8
preshed==1.0.0
promo==1.3.0
prompt-toolkit==1.0.15
protobuf==3.10.0
psutil==5.6.5
ptyprocess==0.5
public==2019.4.13
pyasn1==0.1.9
PyAudio==0.2.8
pycodestyle==2.6.0
pycparser==2.19
pycups==1.9.73
pycurl==7.43.0
pydot==1.2.4
pydub==0.23.1
pyflakes==1.6.0
Pygments==2.2.0
pygobject==3.20.0
pyinotify==0.9.6
pyjsgf==1.8.0
PyJWT==1.7.1
pylint==1.8.4
pyOpenSSL==19.0.0
pyparsing==2.2.0
pypeaks==0.2.7
PyQt5==5.9
pyrsistent==0.15.5
pystaggrelite3==0.1.3
python-apt==1.1.0b1+ubuntu0.16.4.8
python-crfsuite-extension==0.9.7
python-dateutil==2.7.3
python-debian==0.1.27
python-magic==0.4.15
python-mimeparse==0.1.4
python-slugify==4.0.0
python-speech-features==0.6
python-subunit==1.1.0
python-systemd==231
python-xlib==0.14
pytils==0.3
pytz==2018.5
pyvim==0.0.21
pyvttbl==0.5.2.2
pyxdg==0.25
PyYAML==5.1.2
pyzmq==16.0.3
qtconsole==4.3.1
Record==3.4
regex==2017.4.5
reportlab==3.3.0
requests==2.22.0
requests-toolbelt==0.8.0
resampy==0.2.1
researchpy==0.1.7
RMDL==1.0.6
s3transfer==0.1.13
schedule==0.6.0
scikit-learn==0.21.2
scipy==1.1.0
seaborn==0.9.1
sessioninstaller==0.0.0
simplegeneric==0.8.1
simplejson==3.8.1
sip==4.19.8
six==1.12.0
sklearn==0.0
sklearn-pycrfsuite==0.4.0
slip==0.6.1
slip.dbus==0.6.1
smart-open==1.6.0
sortedcontainers==1.5.7
sounddevice==0.3.13
spacy==2.0.5
SQLAlchemy==1.3.11
ssh-import-id==5.5
stanfordcorenlp==3.7.0.2
statistics==1.0.3.5
statsmodels==0.10.1
system-service==0.3
tables==3.2.2
tabulate==0.8.6
tempora==1.9
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-tensorboard==0.1.6
termcolor==1.1.0
terminado==0.8.1
testfixtures==6.14.0
testpath==0.3.1
testscenarios==0.4
testtools==1.8.1
text-unidecode==1.3
TextGrid==1.1
textrazor==1.3.0
Theano==1.0.1
thinc==6.10.2
tools==0.1.9
toolz==0.9.0
tornado==4.5.2
tqdm==4.19.5
traceback2==1.4.0
traitlets==4.3.2
treetaggerwrapper==2.2.4
twine==1.9.1
typing==3.7.4.1
ubuntu-drivers-common==0.0.0
ufw==0.35
ujson==1.35
unattended-upgrades==0.1
Unidecode==0.4.19
unittest2==1.1.0
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.25.6
usb-creator==0.3.0
utils==0.9.0
vboxapi==1.0
virtualenv==16.7.4
virtualenv-clone==0.3.0
VUSentimentLexicon==1.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.1.3
wrapt==1.11.2
xdiagnose==3.8.4.1
xkit==0.0.0
XlsxWriter==0.7.3
zipp==0.6.0
